xbx/CoinAlphaRotationCta中性自适应V3滚动选币实盘框架V3.0/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
anyio==3.3.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
ccxt==1.56.70
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
dynaconf==3.1.7
Flask==2.0.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
pandas==1.1.5
pycares==4.0.0
pycparser==2.20
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sniffio==1.2.0
threadpoolctl==2.2.0
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
